Update clang-tidy to clang version 8
Some new warnings are well intentioned, but not practical to turn on
in GROMACS given the limited suppression capabilities.

Rationales for suppressions of new warnings can be found in
src/.clang-tidy

+++ src/gromacs/awh/pointstate.h
@@ -316,7 +316,7 @@ class PointState
                 return false;
             }
 
-            for (int i = 0; i < numUpdatesSkipped; i++)
+            for (int64_t i = 0; i < numUpdatesSkipped; i++)
             {
                 /* This point was non-local at the time of the update meaning no weight */
                 updateFreeEnergyAndWeight(params, 0, weighthistScaling, logPmfSumScaling);
