+++ src/gromacs/nbnxm/kernels_simd_2xmm/kernel_inner.h
@@ -218,7 +218,7 @@
     ajz           = ajy + STRIDE;
 
 #ifdef CHECK_EXCLS
-    gmx_load_simd_2xnn_interactions(l_cj[cjind].excl,
+    gmx_load_simd_2xnn_interactions(static_cast<int>(l_cj[cjind].excl),
                                     filter_S0, filter_S2,
                                     &interact_S0, &interact_S2);
 #endif /* CHECK_EXCLS */
