+++ src/testutils/testasserts.cpp
@@ -1,7 +1,7 @@
 /*
  * This file is part of the GROMACS molecular simulation package.
  *
- * Copyright (c) 2014,2015,2016,2018, by the GROMACS development team, led by
+ * Copyright (c) 2014,2015,2016,2018,2019, by the GROMACS development team, led by
  * Mark Abraham, David van der Spoel, Berk Hess, and Erik Lindahl,
  * and including many others, as listed in the AUTHORS file in the
  * top-level source directory and at http://www.gromacs.org.
@@ -329,8 +329,8 @@ FloatingPointTolerance
 relativeToleranceAsFloatingPoint(double magnitude, double tolerance)
 {
     const double absoluteTolerance = std::abs(magnitude) * tolerance;
-    return FloatingPointTolerance(absoluteTolerance, absoluteTolerance,
-                                  tolerance, tolerance,
+    return FloatingPointTolerance(static_cast<float>(absoluteTolerance), absoluteTolerance,
+                                  static_cast<float>(tolerance), tolerance,
                                   UINT64_MAX, UINT64_MAX,
                                   false);
 }
